IoT Security dataset Analysis
Performed exploratory data analysis on IoT Security Dataset to identify trends and patterns in the current Attack performed 

Data set Detail:
This dataset contains data collected from a real IoT device, a Raspberry Pi, which
was infected by botnet malware in a controlled environment. The goal of this dataset
is to provide means to observe different perspectives of the operation of the device
such as the use of CPU and memory and the network traffic. The collected data
includes legitimate and malicious behavior . Legitimate data concerns the
emulation of regular behavior of domestic IoT devices, such as surveillance cameras
and media centers . The malicious data is related to botnets activities, including
communication with the Command & Control (C&C) and some DDoS attacks. The
dataset was labeled and the files were separated to organize the different situations
faced during the experiments.